Split admin auth into own listener
